# packages/contracts/generated/openapi/knowthat.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "PAI KnowThat API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/knowthat/query": {
      "post": {
        "operationId": "queryKnowThatV1",
        "security": [
          {
            "PaiWorkloadJwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$id": "urn:pai:knowthat:query-request:v1",
                "anyOf": [
                  {
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "schema_version",
                      "workspace_id",
                      "bot_id",
                      "owner_agent_id",
                      "deployment_environment",
                      "release_channel"
                    ],
                    "properties": {
                      "schema_version": {
                        "const": "knowthat_query.v1",
                        "type": "string"
                      },
                      "workspace_id": {
                        "minLength": 1,
                        "maxLength": 512,
                        "pattern": "^[^\\u0000-\\u001f\\u007f]+$",
                        "type": "string"
                      },
                      "bot_id": {
                        "minLength": 1,
                        "maxLength": 512,
                        "pattern": "^[^\\u0000-\\u001f\\u007f]+$",
                        "type": "string"
                      },
                      "owner_agent_id": {
                        "minLength": 1,
                        "maxLength": 512,
                        "pattern": "^[^\\u0000-\\u001f\\u007f]+$",
                        "type": "string"
                      },
                      "deployment_environment": {
                        "$id": "urn:pai:shared:deployment-environment:v1",
                        "anyOf": [
                          {
                            "const": "local",
                            "type": "string"
                          },
                          {
                            "const": "dev",
                            "type": "string"
                          },
                          {
                            "const": "staging",
                            "type": "string"
                          },
                          {
                            "const": "prod",
                            "type": "string"
                          }
                        ]
                      },
                      "release_channel": {
                        "$id": "urn:pai:shared:release-channel:v1",
                        "anyOf": [
                          {
                            "const": "stable",
                            "type": "string"
                          },
                          {
                            "const": "canary",
                            "type": "string"
                          }
                        ]
                      },
                      "categories": {
                        "maxItems": 7,
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                          "anyOf": [
                            {
                              "const": "user_profile",
                              "type": "string"
                            },
                            {
                              "const": "relationship_state",
                              "type": "string"
                            },
                            {
                              "const": "world_state",
                              "type": "string"
                            },
                            {
                              "const": "self_state",
                              "type": "string"
                            },
                            {
                              "const": "project_fact",
                              "type": "string"
                            },
                            {
                              "const": "preference",
                              "type": "string"
                            },
                            {
                              "const": "rule",
                              "type": "string"
                            }
                          ]
                        }
                      },
                      "query": {
                        "minLength": 1,
                        "maxLength": 4096,
                        "type": "string"
                      },
                      "limit": {
                        "minimum": 1,
                        "maximum": 200,
                        "default": 50,
                        "type": "integer"
                      }
                    }
                  },
                  {
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "schema_version",
                      "workspace_id",
                      "bot_id",
                      "owner_agent_id",
                      "deployment_environment",
                      "release_channel",
                      "snapshot_token",
                      "cursor"
                    ],
                    "properties": {
                      "schema_version": {
                        "const": "knowthat_query.v1",
                        "type": "string"
                      },
                      "workspace_id": {
                        "minLength": 1,
                        "maxLength": 512,
                        "pattern": "^[^\\u0000-\\u001f\\u007f]+$",
                        "type": "string"
                      },
                      "bot_id": {
                        "minLength": 1,
                        "maxLength": 512,
                        "pattern": "^[^\\u0000-\\u001f\\u007f]+$",
                        "type": "string"
                      },
                      "owner_agent_id": {
                        "minLength": 1,
                        "maxLength": 512,
                        "pattern": "^[^\\u0000-\\u001f\\u007f]+$",
                        "type": "string"
                      },
                      "deployment_environment": {
                        "$id": "urn:pai:shared:deployment-environment:v1",
                        "anyOf": [
                          {
                            "const": "local",
                            "type": "string"
                          },
                          {
                            "const": "dev",
                            "type": "string"
                          },
                          {
                            "const": "staging",
                            "type": "string"
                          },
                          {
                            "const": "prod",
                            "type": "string"
                          }
                        ]
                      },
                      "release_channel": {
                        "$id": "urn:pai:shared:release-channel:v1",
                        "anyOf": [
                          {
                            "const": "stable",
                            "type": "string"
                          },
                          {
                            "const": "canary",
                            "type": "string"
                          }
                        ]
                      },
                      "snapshot_token": {
                        "minLength": 16,
                        "maxLength": 8192,
                        "type": "string"
                      },
                      "cursor": {
                        "minLength": 16,
                        "maxLength": 8192,
                        "type": "string"
                      },
                      "limit": {
                        "minimum": 1,
                        "maximum": 200,
                        "default": 50,
                        "type": "integer"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Canonical KnowThat owner response",
            "content": {
              "application/json": {
                "schema": {
                  "$id": "urn:pai:knowthat:query-response:v1",
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "schema_version",
                    "snapshot_token",
                    "query_revision",
                    "as_of",
                    "facts"
                  ],
                  "properties": {
                    "schema_version": {
                      "const": "knowthat_query.v1",
                      "type": "string"
                    },
                    "snapshot_token": {
                      "minLength": 16,
                      "maxLength": 8192,
                      "type": "string"
                    },
                    "query_revision": {
                      "minimum": 1,
                      "maximum": 9007199254740991,
                      "type": "integer"
                    },
                    "as_of": {
                      "minLength": 20,
                      "maxLength": 35,
                      "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d{1,9})?(?:Z|[+-](?:[01]\\d|2[0-3]):[0-5]\\d)$",
                      "type": "string"
                    },
                    "facts": {
                      "maxItems": 200,
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "required": [
                          "id",
                          "semantic_key",
                          "category",
                          "subject",
                          "predicate",
                          "object",
                          "confidence",
                          "updated_at"
                        ],
                        "properties": {
                          "id": {
                            "minLength": 1,
                            "maxLength": 512,
                            "pattern": "^[^\\u0000-\\u001f\\u007f]+$",
                            "type": "string"
                          },
                          "semantic_key": {
                            "minLength": 1,
                            "maxLength": 256,
                            "type": "string"
                          },
                          "category": {
                            "anyOf": [
                              {
                                "const": "user_profile",
                                "type": "string"
                              },
                              {
                                "const": "relationship_state",
                                "type": "string"
                              },
                              {
                                "const": "world_state",
                                "type": "string"
                              },
                              {
                                "const": "self_state",
                                "type": "string"
                              },
                              {
                                "const": "project_fact",
                                "type": "string"
                              },
                              {
                                "const": "preference",
                                "type": "string"
                              },
                              {
                                "const": "rule",
                                "type": "string"
                              }
                            ]
                          },
                          "subject": {
                            "minLength": 1,
                            "maxLength": 4096,
                            "type": "string"
                          },
                          "predicate": {
                            "minLength": 1,
                            "maxLength": 4096,
                            "type": "string"
                          },
                          "object": {
                            "minLength": 1,
                            "maxLength": 4096,
                            "type": "string"
                          },
                          "confidence": {
                            "minimum": 0,
                            "maximum": 1,
                            "type": "number"
                          },
                          "valid_from": {
                            "minLength": 20,
                            "maxLength": 35,
                            "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d{1,9})?(?:Z|[+-](?:[01]\\d|2[0-3]):[0-5]\\d)$",
                            "type": "string"
                          },
                          "valid_until": {
                            "minLength": 20,
                            "maxLength": 35,
                            "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d{1,9})?(?:Z|[+-](?:[01]\\d|2[0-3]):[0-5]\\d)$",
                            "type": "string"
                          },
                          "updated_at": {
                            "minLength": 20,
                            "maxLength": 35,
                            "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12]\\d|3[01])T(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d{1,9})?(?:Z|[+-](?:[01]\\d|2[0-3]):[0-5]\\d)$",
                            "type": "string"
                          }
                        }
                      }
                    },
                    "next_cursor": {
                      "minLength": 16,
                      "maxLength": 8192,
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "authorization_scope_mismatch"
          },
          "404": {
            "description": "owner entity not found in the authorized scope"
          },
          "409": {
            "description": "idempotency, version, or recovery conflict"
          },
          "410": {
            "description": "snapshot_expired"
          },
          "413": {
            "description": "payload_too_large"
          },
          "422": {
            "description": "schema or owner semantic validation failed"
          },
          "503": {
            "description": "owner dependency unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "PaiWorkloadJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}